Updated java build script

+++ clientlauncher/Shared/src/ritzow/sandbox/build/Build.java
@@ -10,7 +10,6 @@ import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.*;
 import java.nio.file.attribute.BasicFileAttributes;
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 import java.util.Locale;
@@ -38,9 +37,6 @@ public class Build {
 		"org.lwjgl.openal"
 	};
 	
-	private static final Path TEMP_SHARED = Path.of("shared");
-	private static final Path TEMP_CLIENT = Path.of("client");
-	
 	private static final List<Path> LWJGL_MODULES = List.of(
 		Path.of("lwjgl.jar"),
 		Path.of("lwjgl-glfw.jar"),
@@ -53,47 +49,56 @@ public class Build {
 		Path CLIENT_DIR = Path.of(args[1]);
 		Path OUTPUT_DIR = Path.of(args[2]);
 
-		Path temp = Files.createTempDirectory(OUTPUT_DIR, "temp");
-		System.out.println("Compiling shared code...");
+		Path temp = Files.createTempDirectory(OUTPUT_DIR, "classes_");
+		Runtime.getRuntime().addShutdownHook(new Thread(() -> {
+			try {
+				delete(temp);
+			} catch (IOException e) {
+				e.printStackTrace();
+			}
+		}));
 		
+		Path sharedOut = temp.resolve("shared");
+		Path sharedSrc = SHARED_DIR.resolve("src");
+		System.out.println("Searching for shared code...");
+		var sharedFiles = getSourceFiles(sharedSrc);
+		System.out.println("Compiling shared code...");
 		boolean sharedSuccess = compile(
-			temp.resolve(TEMP_SHARED), 
-			SHARED_DIR, 
-			getSourceFiles(SHARED_DIR.resolve("src")), 
+			sharedOut, 
+			sharedSrc, 
+			sharedFiles, 
 			List.of()
 		);
 		
 		if(sharedSuccess) {
 			System.out.println("Shared code compiled.");
-			System.out.println("Compiling client code...");
-			
-			Collection<Path> modules = new ArrayList<Path>();
-			
-			modules.add(temp.resolve(TEMP_SHARED));
-			
-			LWJGL_MODULES.stream()
+			Collection<Path> modules = LWJGL_MODULES.stream()
 				.map(CLIENT_DIR.resolve("libraries").resolve("lwjgl")::resolve)
-				.forEach(modules::add);
-			
+				.collect(Collectors.toList());
+			modules.add(sharedOut);
 			Path clientSrc = CLIENT_DIR.resolve("src");
-			
+			Path clientOut = temp.resolve("client");
+			System.out.println("Searching for client code...");
+			var clientFiles = getSourceFiles(clientSrc);
+			System.out.println("Compiling client code...");
 			boolean clientSuccess = compile(
-				temp.resolve(TEMP_CLIENT), 
+				clientOut, 
 				clientSrc,
-				getSourceFiles(clientSrc), 
+				clientFiles, 
 				modules
 			);
 			
 			if(clientSuccess) {
 				System.out.print("Client code compiled.\nRunning jlink...");
-				modules.add(temp.resolve(TEMP_CLIENT));
+				modules.add(clientOut);
 				int result = jlink(OUTPUT_DIR.resolve("jvm"), MODULE_NAMES, modules);
 				System.out.println(result == 0 ? "jlink successful." : "jlink failed with error " + result + ".");
 			} else {
-				System.out.println("Compilation failed.");
+				System.out.println("Client compilation failed.");
 			}
+		} else {
+			System.out.println("Shared compilation failed.");
 		}
-		delete(temp);
 	}
 	
 	private static Iterable<JavaFileObject> getSourceFiles(Path src) throws IOException {
@@ -158,6 +163,8 @@ public class Build {
 			
 			@Override
 			public FileVisitResult postVisitDirectory(Path dir, IOException exc) throws IOException {
+		        if (exc != null)
+		            throw exc;
 				Files.delete(dir);
 				return FileVisitResult.CONTINUE;
 			}
